requirements.txt
flask
line-bot-sdk
google-generativeai
Pillow
python-dotenv